perf(client): drop extra file
